Working ProductPriceCalculator

+++ src/main/java/org/pcdv/valocb/product/Product.java
@@ -1,11 +1,15 @@
 package org.pcdv.valocb.product;
 
-
 import org.pcdv.valocb.currency.CurrencyCode;
+
+import java.math.BigDecimal;
 
 public interface Product {
     String getName();
+
     String getUnderlying();
+
     CurrencyCode getCurrencyCode();
-    int getPrice();
+
+    BigDecimal getPrice();
 }
