+++ src/main/java/org/pcdv/valocb/product/ProductImpl.java
@@ -3,10 +3,12 @@ package org.pcdv.valocb.product;
 import lombok.Data;
 import org.pcdv.valocb.currency.CurrencyCode;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductImpl implements Product {
     private final String name;
     private final String underlying;
     private final CurrencyCode currencyCode;
-    private final int price;
+    private final BigDecimal price;
 }
